# apps/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SummerGear Go API",
    "version": "0.6.0",
    "description": "Go service-session authentication, listings, and private listing image lifecycle. Social provider credentials and Storage credentials may remain disabled in isolated test environments."
  },
  "paths": {
...
    "/api/v1/orders": {
      "post": {
        "operationId": "createOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created with inventory reservation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or malformed UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "CSRF, origin or ownership rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order or purchasable listing not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stock or order state conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage or payment provider unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "description": "Single server listing, quantity 1. Repeating a pending purchase for the same buyer and listing returns the same order. Stock must be explicitly provisioned for an approved seller."
      },
      "get": {
        "operationId": "listOrders",
        "responses": {
          "200": {
            "description": "Buyer's orders.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or malformed UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "CSRF, origin or ownership rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order or purchasable listing not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stock or order state conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage or payment provider unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/orders/{orderId}": {
      "get": {
        "operationId": "getOrder",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderID"
          }
        ],
        "responses": {
          "200": {
            "description": "Order detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or malformed UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "CSRF, origin or ownership rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order or purchasable listing not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stock or order state conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage or payment provider unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/orders/{orderId}/cancel": {
      "post": {
        "operationId": "cancelOrder",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderID"
          }
        ],
        "responses": {
          "200": {
            "description": "Unpaid order cancelled; duplicate cancellation does not release stock twice.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or malformed UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "CSRF, origin or ownership rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order or purchasable listing not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stock or order state conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage or payment provider unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/orders/{orderId}/payments/confirm": {
      "post": {
        "operationId": "confirmPayment",
        "description": "Test-only approval using configured Toss widget test credentials, or the isolated fake fixture gateway when keys are absent. Server validates buyer, amount and provider. Pending outcomes require authoritative lookup; replays do not approve again.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server verified payment result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "202": {
            "description": "Approval outcome is pending; poll order status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or malformed UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "CSRF, origin or ownership rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order or purchasable listing not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stock or order state conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage or payment provider unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{orderId}/refunds": {
      "post": {
        "operationId": "refundOrder",
        "description": "Buyer-only full test cancellation before handover. CSRF required. One durable cancellation per order; same provider idempotency key on retry. No partial refunds or seller payouts.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server verified payment result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "202": {
            "description": "Cancellation outcome pending; poll order status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or malformed UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "CSRF, origin or ownership rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order or purchasable listing not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stock or order state conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage or payment provider unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{orderId}/payments/config": {
      "get": {
        "operationId": "getPaymentConfig",
        "description": "Buyer-only public SDK configuration. No secret key is returned.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderID"
          }
        ],
        "responses": {
          "200": {
            "description": "Public test payment configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "provider",
                    "clientKey",
                    "customerKey"
                  ],
                  "properties": {
                    "provider": {
                      "type": "string",
                      "enum": [
                        "fake_toss",
                        "toss_test"
                      ]
                    },
                    "clientKey": {
                      "type": "string"
                    },
                    "customerKey": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or malformed UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "CSRF, origin or ownership rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order or purchasable listing not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Stock or order state conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage or payment provider unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/seller/orders": { "get": { "operationId": "listSellerOrders", "description": "Latest 50 orders of current member's active seller memberships; payment approval is distinct from handover.", "responses": { "200": { "description": "Seller orders.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderListResponse" } } } }, "401": { "$ref": "#/components/responses/SessionRequired" } } } },
    "/api/v1/seller/orders/{id}/accept": { "post": { "operationId": "acceptSellerOrder", "description": "Active seller membership transitions approved, confirmed awaiting_acceptance order to accepted. Replays are idempotent.", "parameters": [{ "$ref": "#/components/parameters/ListingID" }, { "$ref": "#/components/parameters/CSRFToken" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false } } } }, "responses": { "200": { "description": "Accepted order.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderView" } } } }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } } } },
    "/api/v1/seller/orders/{id}/hand-over": { "post": { "operationId": "handOverSellerOrder", "description": "Active seller transitions accepted order to handed_over; full buyer refund is no longer permitted.", "parameters": [{ "$ref": "#/components/parameters/ListingID" }, { "$ref": "#/components/parameters/CSRFToken" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false } } } }, "responses": { "200": { "description": "Handed-over order.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderView" } } } }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } } } },
    "/api/v1/orders/{orderId}/receive": { "post": { "operationId": "receiveOrder", "description": "Buyer confirms receipt after handover; this is completion, not payment approval.", "parameters": [{ "$ref": "#/components/parameters/OrderID" }, { "$ref": "#/components/parameters/CSRFToken" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false } } } }, "responses": { "200": { "description": "Completed order.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderView" } } } }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } } } },
    "/api/v1/orders/{orderId}/fulfillment": { "get": { "operationId": "orderFulfillmentHistory", "description": "Buyer or active seller membership sees audited fulfillment transitions.", "parameters": [{ "$ref": "#/components/parameters/OrderID" }], "responses": { "200": { "description": "Fulfillment event history.", "content": { "application/json": { "schema": { "type": "object", "required": ["events"], "properties": { "events": { "type": "array", "items": { "$ref": "#/components/schemas/FulfillmentEvent" } } } } } } }, "404": { "$ref": "#/components/responses/NotFound" } } } },
...
  },
  "components": {
    "responses": {
...
      "SessionRequired": { "description": "Session required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "NotFound": { "description": "Listing unavailable to requester.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "State conflict.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "parameters": {
...
      "ListingID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "CSRFToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "OrderID": {
        "name": "orderId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "schemas": {
...
      "FulfillmentEvent": { "type": "object", "additionalProperties": false, "required": ["id", "actorMemberId", "fromStatus", "toStatus", "createdAt"], "properties": { "id": { "type": "integer" }, "actorMemberId": { "type": "string", "format": "uuid" }, "fromStatus": { "type": "string", "enum": ["awaiting_acceptance", "accepted", "handed_over"] }, "toStatus": { "type": "string", "enum": ["accepted", "handed_over", "completed"] }, "createdAt": { "type": "string", "format": "date-time" } } },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "pattern": "^[a-f0-9]{32}$"
          }
        }
      },
...
      "CreateOrderRequest": {
        "type": "object",
        "required": [
          "listingId",
          "quantity"
        ],
        "additionalProperties": false,
        "properties": {
          "listingId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "const": 1
          }
        }
      },
      "OrderView": {
        "type": "object",
        "required": [
          "id",
          "buyerId",
          "sellerId",
          "listingId",
          "itemName",
          "unitPriceKrw",
          "quantity",
          "totalAmountKrw",
          "currency",
          "status",
          "paymentStatus",
          "fulfillmentStatus",
          "acceptedAt",
          "handedOverAt",
          "receivedAt",
          "createdAt",
          "shippingFeeKrw",
          "serviceFeeKrw",
          "updatedAt"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "buyerId": {
            "type": "string",
            "format": "uuid"
          },
          "sellerId": {
            "type": "string",
            "format": "uuid"
          },
          "listingId": {
            "type": "string",
            "format": "uuid"
          },
          "itemName": {
            "type": "string"
          },
          "unitPriceKrw": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer",
            "const": 1
          },
          "shippingFeeKrw": {
            "type": "integer"
          },
          "serviceFeeKrw": {
            "type": "integer"
          },
          "totalAmountKrw": {
            "type": "integer"
          },
          "currency": {
            "type": "string",
            "enum": [
              "KRW"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "cancelled",
              "confirmed"
            ]
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "unpaid",
              "pending_approval",
              "approved",
              "failed",
              "pending_cancel",
              "cancelled"
            ]
          },
          "fulfillmentStatus": { "type": "string", "enum": ["awaiting_acceptance", "accepted", "handed_over", "completed"] },
          "acceptedAt": { "type": ["string", "null"], "format": "date-time" },
          "handedOverAt": { "type": ["string", "null"], "format": "date-time" },
          "receivedAt": { "type": ["string", "null"], "format": "date-time" },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "reservation": {
            "$ref": "#/components/schemas/Reservation"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "OrderListResponse": {
        "type": "object",
        "required": [
          "orders"
        ],
        "additionalProperties": false,
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderView"
            }
          }
        }
      },
      "CancelOrderResponse": {
        "type": "object",
        "required": [
          "order"
        ],
        "additionalProperties": false,
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderView"
          }
        }
      },
      "ConfirmPaymentRequest": {
        "type": "object",
        "required": [
          "paymentKey",
          "amount"
        ],
        "additionalProperties": false,
        "properties": {
          "paymentKey": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,200}$"
          },
          "amount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 999999999999
          }
        }
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-sg_session"
      }
    }
  }
}